sw_services:xilasu:Populated Nonce field is inaccurate
The nonce buffer was declared as a u32 and populated byte-by-byte, which resulted in an incorrect nonce value.
The declaration has now been updated to a u8 type to ensure the nonce is handled correctly.

Acked-by: Mounika Akula <mounika.akula@amd.com>

+++ lib/sw_services/xilasu/examples/xasu_ocp_dme_example.c
@@ -80,7 +80,7 @@ static void XAsu_OcpDmePrintData(const u8 *Data, u32 DataLen, const char *BufNam
 static u8 Notify = 0;			/**< To notify the call back from client library */
 static u32 ErrorStatus = XST_FAILURE;	/**< Variable holds the status of the OCP-DME operation from
 					client library and gets updated during callback. */
-static u32 NonceBuff[XASU_OCP_DME_NONCE_SIZE_IN_BYTES] __attribute__ ((section (".data.Nonce"))) =
+static u8 NonceBuff[XASU_OCP_DME_NONCE_SIZE_IN_BYTES] __attribute__ ((section (".data.Nonce"))) =
 {
 	0x70,0x69,0x77,0x35,0x0b,0x93,
 	0x92,0xa0,0x48,0x2c,0xd8,0x23,
